refactor components and functional firts attemp to create new search
